+++ src/api/events/index.ts
@@ -8,10 +8,10 @@ import type {
 	ApplicationDidTerminate,
 	DidReceiveDeepLink,
 	DidReceiveGlobalSettings,
+	DidReceiveSecrets,
 	SystemDidWakeUp,
 } from "./system";
 import type {
-	DidReceivePluginMessage,
 	DidReceivePropertyInspectorMessage,
 	PropertyInspectorDidAppear,
 	PropertyInspectorDidDisappear,
@@ -39,7 +39,6 @@ export {
 	type SystemDidWakeUp,
 } from "./system";
 export {
-	type DidReceivePluginMessage,
 	type DidReceivePropertyInspectorMessage,
 	type PropertyInspectorDidAppear,
 	type PropertyInspectorDidDisappear,
@@ -70,6 +69,7 @@ export type PluginEvent =
 	| DidReceiveDeepLink
 	| DidReceiveGlobalSettings<JsonObject>
 	| DidReceivePropertyInspectorMessage<JsonValue>
+	| DidReceiveSecrets<JsonObject>
 	| DidReceiveSettings<JsonObject>
 	| KeyDown<JsonObject>
 	| KeyUp<JsonObject>
@@ -86,19 +86,4 @@ export type PluginEvent =
  */
 export type PluginEventMap = {
 	[K in PluginEvent["event"]]: [event: Extract<PluginEvent, EventIdentifier<K>>];
-};
-
-/**
- * Events received by the UI, from Stream Deck.
- */
-export type UIEvent =
-	| DidReceiveGlobalSettings<JsonObject>
-	| DidReceivePluginMessage<JsonValue>
-	| DidReceiveSettings<JsonObject>;
-
-/**
- * Map of events received by the UI, from Stream Deck.
- */
-export type UIEventMap = {
-	[K in UIEvent["event"]]: [event: Extract<UIEvent, EventIdentifier<K>>];
 };